Add super kernel in moe

+++ vllm_ascend/models/deepseek_v2.py
@@ -246,7 +246,7 @@ class CustomDeepseekV2MoE(nn.Module):
                                      prefix=f"{prefix}.gate")
         if config.topk_method == "noaux_tc":
             self.gate.e_score_correction_bias = nn.Parameter(
-                torch.empty(config.n_routed_experts))
+                torch.empty(config.n_routed_experts, dtype=torch.float))
         else:
             self.gate.e_score_correction_bias = None
